+++ library/ruler/src/main/java/io/mehow/ruler/ImperialLengthUnit.kt
@@ -25,8 +25,8 @@ enum class ImperialLengthUnit(
     return Distance.create(value.toBigDecimal() * meterRatio)
   }
 
-  override fun toMeasuredLength(distance: BigDecimal): Double {
-    return distance.divide(meterRatio, 9, DOWN).toDouble()
+  override fun toMeasuredLength(distance: BigDecimal): BigDecimal {
+    return distance.divide(meterRatio, 9, DOWN)
   }
 
   override fun appliesRangeTo(distance: BigDecimal): Boolean {
